fix(ci): rename motion-tokens to motion-presets
The previous commit added `packages/frontend/src/lib/motion-tokens.ts`
locally and `git add -A` silently dropped it because the author's
global gitignore had a `*token*` glob. The PR landed with seven
files importing from a path that didn't exist in-tree, so CI's
`tsc` step crashed on every consumer.

Renamed the module to `motion-presets.ts` so the import path is
greppable from the build output and not subject to a global-ignore
trap. Rewrote all seven import sites; no behavioral change.
